Fix Docx Parser & Docx Translate
